+++ bmc-common/src/test/java/com/oracle/bmc/http/signing/internal/PEMFileRSAPrivateKeySupplierTest.java
@@ -1,0 +1,24 @@
+package com.oracle.bmc.http.signing.internal;
+
+import org.junit.Test;
+
+import java.io.ByteArrayInputStream;
+import java.io.InputStream;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+
+import static org.junit.Assert.fail;
+
+/**
+ * Tests for {@link PEMFileRSAPrivateKeySupplierTest}.
+ */
+public class PEMFileRSAPrivateKeySupplierTest {
+    @Test(expected=IllegalArgumentException.class)
+    public void ctor_invalidFile() throws IOException {
+        InputStream notAPem = new ByteArrayInputStream(new byte[0]);
+        // not a valid key file
+        new PEMFileRSAPrivateKeySupplier(notAPem, null);
+        fail("Should have thrown an IllegalArgumentException");
+    }
+}